backend/pyproject.toml
[project]
name = "souksync-backend"
version = "0.1.0"
description = "SoukSync FastAPI backend"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "pydantic-settings>=2.2",
    "structlog>=24.1",
    "redis>=5.0",
    "alembic>=1.13",
    "celery>=5.3",
    "httpx>=0.27",
    "pyjwt>=2.8",
    "eval-type-backport>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "ruff>=0.4",
    "mypy>=1.9",
    "aiosqlite>=0.20",
]

[tool.ruff]
target-version = "py39"
line-length = 120
src = ["app", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "TCH"]
# FastAPI uses Depends() in default args intentionally
ignore = ["B008"]

[tool.mypy]
python_version = "3.9"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
